Refactor CalloutAlert
