+++ src/containers/SceaHomepageSpeciesContainer.js
@@ -3,28 +3,8 @@ import PropTypes from 'prop-types'
 import URI from 'urijs'
 import styled from 'styled-components'
 
+import CalloutAlert from './CalloutAlert'
 import SpeciesCard from '../cards/SpeciesCard'
-
-const CalloutAlert = ({error}) =>
-  <div className={`row column`}>
-    <div className={`callout alert small`}>
-      <h5>Oops!</h5>
-      <p>
-        {error.description}<br/>
-        If the error persists, in order to help us debug the issue, please copy the URL and this message and
-        send it to us via <a href={`https://www.ebi.ac.uk/support/gxasc`}>the EBI Support & Feedback system</a>:
-      </p>
-      <code>{`${error.name}: ${error.message}`}</code>
-    </div>
-  </div>
-
-CalloutAlert.propTypes = {
-  error: PropTypes.shape({
-    description: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-    message: PropTypes.string.isRequired
-  })
-}
 
 const CardContainer = styled.div`
   border-radius: 8px;
